+++ sample-core/src/main/kotlin/com/alecstrong/sql/psi/sample/core/SampleFile.kt
@@ -6,7 +6,7 @@ import com.intellij.psi.FileViewProvider
 class SampleFile(viewProvider: FileViewProvider) : SqlFileBase(viewProvider, SampleLanguage, predefinedTables = emptyList()) {
   override val order = name.substringBefore(".${fileType.defaultExtension}").let { name ->
     if (name.all { it in '0'..'9' }) {
-      name.toInt()
+      name.toLong()
     } else {
       null
     }
